no need for static_cast

+++ contrib/mul/vil3d/vil3d_convert.h
@@ -1,4 +1,4 @@
-// This is contrib/mul/vil3d/vil3d_convert.h
+// This is mul/vil3d/vil3d_convert.h
 #ifndef vil3d_convert_h_
 #define vil3d_convert_h_
 #ifdef VCL_NEEDS_PRAGMA_INTERFACE
@@ -18,14 +18,14 @@
 // vil3d_image_view_base_sptr. The basic conversion
 // operations (e.g. casting, rounding) are available in both types.
 // All of the conversions attempt to find shortcuts, so the output
-// may be a reconfigured, or shallow copy of the input. 
+// may be a reconfigured, or shallow copy of the input.
 //
 // \par vil3d_convert with explicit pixel types
 // These are useful when you have two vil3d_image_view objects you want
 // to convert between. You can use them in templates where the pixel
 // type is one of the template parameters. These functions
 // may create a shallow copy of the input if the types match too save
-// unnecessary work. 
+// unnecessary work.
 // - vil3d_convert_cast
 // - vil3d_convert_round
 // - vil3d_convert_stretch_range
@@ -35,7 +35,7 @@
 // of vil3d_image_resource::get_view(), where you may not know the
 // pixel type in advance, but want to force the image into a
 // particular pixel type.
-// - vil3d_convert_cast 
+// - vil3d_convert_cast
 // - vil3d_convert_round
 // - vil3d_convert_stretch_range
 // - vil3d_convert_to_n_planes
@@ -48,14 +48,13 @@
 // Note that these vil3d_convert_..( vil3d_image_view_base_sptr ) functions
 // are provided as a convenience for users of vil3d_load and
 // vil3d_image_resource::get_view(). Their existence should not suggest
-// that it is sensible to use a vil3d_image_view_base_sptr as storage, 
+// that it is sensible to use a vil3d_image_view_base_sptr as storage,
 // nor that it is a good idea to write a functions that
 // take or return a vil3d_image_view_base_sptr. If you need a
 // pixel-type-agnostic image container then use a vil3d_image_resource_sptr
 
 // It may be a good idea to provide vil3d_image_resource_sptr based
 // vil3d_converts as well.
-
 
 #include <vcl_cassert.h>
 #include <vcl_limits.h>
@@ -66,8 +65,6 @@
 #include <vil3d/vil3d_copy.h>
 
 
-
-
 //: Cast one pixel type to another (with rounding).
 // There must be a cast operator from inP to outP
 //
@@ -76,7 +73,7 @@
 // \relates vil3d_image_view
 template <class inP, class outP>
 inline void vil3d_convert_cast(const vil3d_image_view<inP >&src,
-  vil3d_image_view<outP >&dest)
+                               vil3d_image_view<outP >&dest)
 {
   if (vil_pixel_format_of(inP()) == vil_pixel_format_of(outP()))
     dest = src;
@@ -95,7 +92,7 @@ inline void vil3d_convert_cast(const vil3d_image_view<inP >&src,
 // \relates vil3d_image_view
 template <class inP, class outP>
 inline void vil3d_convert_round(const vil3d_image_view<inP >&src,
-  vil3d_image_view<outP >&dest)
+                                vil3d_image_view<outP >&dest)
 {
   if (vil_pixel_format_of(inP()) == vil_pixel_format_of(outP()))
     dest = src;
@@ -104,12 +101,11 @@ inline void vil3d_convert_round(const vil3d_image_view<inP >&src,
 }
 
 
-
 //: Convert src to byte image dest by stretching to range [0,255]
 // \relates vil3d_image_view
 template <class T>
 inline void vil3d_convert_stretch_range(const vil3d_image_view<T>& src,
-                                      vil3d_image_view<vxl_byte>& dest)
+                                        vil3d_image_view<vxl_byte>& dest)
 {
   T min_b,max_b;
   vil3d_math_value_range(src,min_b,max_b);
@@ -135,15 +131,15 @@ inline void vil3d_convert_stretch_range(const vil3d_image_view<T>& src,
 //: Convert src to double image dest by stretching to range [lo,hi]
 template <class inP>
 inline void vil3d_convert_stretch_range(const vil3d_image_view<inP>& src,
-                                      vil3d_image_view<double>& dest,
-                                      double lo, double hi)
+                                        vil3d_image_view<double>& dest,
+                                        double lo, double hi)
 {
   inP min_b=0, max_b=0;
   vil3d_math_value_range(src,min_b,max_b);
   double b = 0.0;
   if (max_b-min_b >0)
-    b = static_cast<double>(hi-lo)/static_cast<double>(max_b-min_b);
-  double a = -1.0*min_b*b + lo;    
+    b = (hi-lo)/static_cast<double>(max_b-min_b);
+  double a = -1.0*min_b*b + lo;
   dest.set_size(src.ni(), src.nj(), src.nk(), src.nplanes());
   for (unsigned p = 0; p < src.nplanes(); ++p)
     for (unsigned k = 0; k < src.nk(); ++k)
@@ -151,8 +147,6 @@ inline void vil3d_convert_stretch_range(const vil3d_image_view<inP>& src,
         for (unsigned i = 0; i < src.ni(); ++i)
           dest(i,j,k,p) =  b*src(i,j,k,p) + a;
 }
-
-
 
 
 //: Cast the unknown pixel type to the known one.
@@ -166,20 +160,20 @@ inline void vil3d_convert_stretch_range(const vil3d_image_view<inP>& src,
 // The input image's storage arrangement may not be preserved.
 template <class outP>
 inline vil3d_image_view_base_sptr vil3d_convert_cast(outP /*dummy*/,
-  const vil3d_image_view_base_sptr& src)
+                                                     const vil3d_image_view_base_sptr& src)
 {
   if (!src) return vil3d_image_view_base_sptr();
-  
+
   vil3d_image_view_base_sptr dest = new vil3d_image_view<outP>;
   vil3d_image_view<outP> & dest_ref = static_cast<vil3d_image_view<outP> &>(*dest);
-  
+
   switch( vil_pixel_format_component_format(src->pixel_format()) )
   {
 #define macro(F , T) \
    case F: \
      vil3d_convert_cast( vil3d_image_view<T >( src ), dest_ref );\
      break;
-    
+
     macro( VIL_PIXEL_FORMAT_UINT_32, vxl_uint_32 )
     macro( VIL_PIXEL_FORMAT_INT_32, vxl_int_32 )
     macro( VIL_PIXEL_FORMAT_UINT_16, vxl_uint_16 )
@@ -197,8 +191,6 @@ inline vil3d_image_view_base_sptr vil3d_convert_cast(outP /*dummy*/,
 }
 
 
-
-
 //: Convert an image of any pixel type to another with rounding.
 // This should only be used to convert to scalar
 // pixel types. This function only rounds in terms of the
@@ -209,7 +201,7 @@ inline vil3d_image_view_base_sptr vil3d_convert_cast(outP /*dummy*/,
 //
 // If the input image already has outP as its pixel type, the destination
 // may only be a shallow copy of the source.
-// outP should be a scalar pixel type. 
+// outP should be a scalar pixel type.
 // The input image's storage arrangement may not be preserved.
 template <class outP>
 inline vil3d_image_view_base_sptr vil3d_convert_round(
@@ -222,32 +214,30 @@ inline vil3d_image_view_base_sptr vil3d_convert_round(
   if (vil_pixel_format_component_format(src->pixel_format()) ==
       vil_pixel_format_of(outP()))
     return src;
-    
+
   vil3d_image_view_base_sptr dest = new vil3d_image_view<outP >;
   vil3d_image_view<outP > &dest_ref = static_cast<vil3d_image_view<outP >&>(*dest);
 
   switch(vil_pixel_format_component_format(src->pixel_format()))
   {
 #define macro( F , T ) \
-  case F: { \
+   case F: { \
     vil3d_image_view<T > src1 = src; \
     vil3d_transform2(src1, dest_ref, vil_convert_round_pixel<T , outP>()); \
     break; }
-macro(VIL_PIXEL_FORMAT_BYTE, vxl_byte )
-macro(VIL_PIXEL_FORMAT_SBYTE , vxl_sbyte )
-macro(VIL_PIXEL_FORMAT_UINT_32 , vxl_uint_32 )
-macro(VIL_PIXEL_FORMAT_UINT_16 , vxl_uint_16 )
-macro(VIL_PIXEL_FORMAT_INT_32 , vxl_int_32 )
-macro(VIL_PIXEL_FORMAT_INT_16 , vxl_int_16 )
-macro(VIL_PIXEL_FORMAT_FLOAT , float )
-macro(VIL_PIXEL_FORMAT_DOUBLE , double )
+    macro(VIL_PIXEL_FORMAT_BYTE, vxl_byte )
+    macro(VIL_PIXEL_FORMAT_SBYTE , vxl_sbyte )
+    macro(VIL_PIXEL_FORMAT_UINT_32 , vxl_uint_32 )
+    macro(VIL_PIXEL_FORMAT_UINT_16 , vxl_uint_16 )
+    macro(VIL_PIXEL_FORMAT_INT_32 , vxl_int_32 )
+    macro(VIL_PIXEL_FORMAT_INT_16 , vxl_int_16 )
+    macro(VIL_PIXEL_FORMAT_FLOAT , float )
+    macro(VIL_PIXEL_FORMAT_DOUBLE , double )
 #undef macro
-  default: dest=0;
+   default: dest=0;
   }
   return dest;
 }
-
-
 
 
 //: Create a greyscale image of specified pixel type from any image src.
@@ -269,26 +259,26 @@ inline vil3d_image_view_base_sptr vil3d_convert_to_grey_using_average(
   switch (vil_pixel_format_component_format(src->pixel_format()))
   {
 #define macro( F , T ) \
-  case F: { \
-      /* try to do it quickly */ \
-      if (src->nplanes() == 1 && \
-          vil_pixel_format_component_format(src->pixel_format())==1) \
-        return src; \
-      /* create output view */ \
-      vil3d_image_view<T > dest; \
-      vil3d_image_view<T > src1 = *src; \
-      vil3d_math_mean_over_planes(src1, dest, double()); \
-      return vil3d_image_view_base_sptr(new vil3d_image_view<T >(dest)); }
-macro(VIL_PIXEL_FORMAT_BYTE, vxl_byte )
-macro(VIL_PIXEL_FORMAT_SBYTE , vxl_sbyte )
-macro(VIL_PIXEL_FORMAT_UINT_32 , vxl_uint_32 )
-macro(VIL_PIXEL_FORMAT_UINT_16 , vxl_uint_16 )
-macro(VIL_PIXEL_FORMAT_INT_32 , vxl_int_32 )
-macro(VIL_PIXEL_FORMAT_INT_16 , vxl_int_16 )
-macro(VIL_PIXEL_FORMAT_FLOAT , float )
-macro(VIL_PIXEL_FORMAT_DOUBLE , double )
+   case F: { \
+    /* try to do it quickly */ \
+    if (src->nplanes() == 1 && \
+        vil_pixel_format_component_format(src->pixel_format())==1) \
+      return src; \
+    /* create output view */ \
+    vil3d_image_view<T > dest; \
+    vil3d_image_view<T > src1 = *src; \
+    vil3d_math_mean_over_planes(src1, dest, double()); \
+    return vil3d_image_view_base_sptr(new vil3d_image_view<T >(dest)); }
+    macro(VIL_PIXEL_FORMAT_BYTE, vxl_byte )
+    macro(VIL_PIXEL_FORMAT_SBYTE , vxl_sbyte )
+    macro(VIL_PIXEL_FORMAT_UINT_32 , vxl_uint_32 )
+    macro(VIL_PIXEL_FORMAT_UINT_16 , vxl_uint_16 )
+    macro(VIL_PIXEL_FORMAT_INT_32 , vxl_int_32 )
+    macro(VIL_PIXEL_FORMAT_INT_16 , vxl_int_16 )
+    macro(VIL_PIXEL_FORMAT_FLOAT , float )
+    macro(VIL_PIXEL_FORMAT_DOUBLE , double )
 #undef macro
-  default:
+   default:
     return vil3d_image_view_base_sptr();
   }
 }
@@ -315,39 +305,38 @@ inline vil3d_image_view_base_sptr vil3d_convert_to_n_planes(
   switch (vil_pixel_format_component_format(src->pixel_format()))
   {
  #define macro( F, T ) \
-    case F: { \
-        vil3d_image_view<T > src_ref = src; \
-        if (!src_ref) return vil3d_image_view_base_sptr(); \
-        /* try to do it quickly 1 */ \
-        if (src_ref.nplanes() >= n_planes)  /* reduce number of planes */ \
-          return vil3d_image_view_base_sptr( new vil3d_image_view<T >( \
-              vil3d_planes(vil3d_image_view<T > (src),0,1,n_planes) )); \
-        else { /* expand number of planes with copying */ \
-          vil3d_image_view_base_sptr dest = new vil3d_image_view<T >( \
-            src_ref.ni(), src_ref.nj(), src_ref.nk(), n_planes); \
-          vil3d_image_view<T > & dest_ref = \
-            static_cast<vil3d_image_view<T > &>(*dest); \
-          vil3d_image_view<T > dest_slices = \
-            vil3d_planes(dest_ref, 0, 1, src_ref.nplanes()); \
-          vil3d_copy_reformat(src_ref, dest_slices); \
-          vil3d_image_view<T > src_slice(vil3d_plane(src_ref, 0)); \
-          for (unsigned i=src_ref.nplanes(); i<n_planes; ++i) { \
-            dest_slices = vil3d_plane(dest_ref, i); \
-            vil3d_copy_reformat(src_slice,  dest_slices); } \
-          return dest;  } } \
+   case F: { \
+    vil3d_image_view<T > src_ref = src; \
+    if (!src_ref) return vil3d_image_view_base_sptr(); \
+    /* try to do it quickly 1 */ \
+    if (src_ref.nplanes() >= n_planes)  /* reduce number of planes */ \
+      return vil3d_image_view_base_sptr( new vil3d_image_view<T >( \
+          vil3d_planes(vil3d_image_view<T > (src),0,1,n_planes) )); \
+    else { /* expand number of planes with copying */ \
+      vil3d_image_view_base_sptr dest = new vil3d_image_view<T >( \
+        src_ref.ni(), src_ref.nj(), src_ref.nk(), n_planes); \
+      vil3d_image_view<T > & dest_ref = \
+        static_cast<vil3d_image_view<T > &>(*dest); \
+      vil3d_image_view<T > dest_slices = \
+        vil3d_planes(dest_ref, 0, 1, src_ref.nplanes()); \
+      vil3d_copy_reformat(src_ref, dest_slices); \
+      vil3d_image_view<T > src_slice(vil3d_plane(src_ref, 0)); \
+      for (unsigned i=src_ref.nplanes(); i<n_planes; ++i) { \
+        dest_slices = vil3d_plane(dest_ref, i); \
+        vil3d_copy_reformat(src_slice,  dest_slices); } \
+      return dest;  } } \
 
-macro(VIL_PIXEL_FORMAT_BYTE, vxl_byte )
-macro(VIL_PIXEL_FORMAT_SBYTE , vxl_sbyte )
-macro(VIL_PIXEL_FORMAT_UINT_32 , vxl_uint_32 )
-macro(VIL_PIXEL_FORMAT_UINT_16 , vxl_uint_16 )
-macro(VIL_PIXEL_FORMAT_INT_32 , vxl_int_32 )
-macro(VIL_PIXEL_FORMAT_INT_16 , vxl_int_16 )
-macro(VIL_PIXEL_FORMAT_FLOAT , float )
-macro(VIL_PIXEL_FORMAT_DOUBLE , double )
+    macro(VIL_PIXEL_FORMAT_BYTE, vxl_byte )
+    macro(VIL_PIXEL_FORMAT_SBYTE , vxl_sbyte )
+    macro(VIL_PIXEL_FORMAT_UINT_32 , vxl_uint_32 )
+    macro(VIL_PIXEL_FORMAT_UINT_16 , vxl_uint_16 )
+    macro(VIL_PIXEL_FORMAT_INT_32 , vxl_int_32 )
+    macro(VIL_PIXEL_FORMAT_INT_16 , vxl_int_16 )
+    macro(VIL_PIXEL_FORMAT_FLOAT , float )
+    macro(VIL_PIXEL_FORMAT_DOUBLE , double )
 #undef macro
 
-
-  default:
+   default:
     return vil3d_image_view_base_sptr();
   }
 }
@@ -373,7 +362,7 @@ inline vil3d_image_view_base_sptr vil3d_convert_stretch_range(
     return vil3d_image_view_base_sptr();
 
   double hi,lo;
-  
+
   if (vcl_numeric_limits<outP>::is_integer)
   {
     hi = vcl_numeric_limits<outP>::max()+0.999;
@@ -383,7 +372,7 @@ inline vil3d_image_view_base_sptr vil3d_convert_stretch_range(
   {
     hi=1.0;
     lo=0.0;
-  }  
+  }
 
   vil3d_image_view_base_sptr dest = new vil3d_image_view<outP>;
   vil3d_image_view<outP> & dest_ref = static_cast<vil3d_image_view<outP> &>(*dest);
@@ -391,23 +380,23 @@ inline vil3d_image_view_base_sptr vil3d_convert_stretch_range(
   switch (vil_pixel_format_component_format(src->pixel_format()))
   {
 #define macro( F , T ) \
-  case F: { \
-      vil3d_image_view<T> src_ref = src; \
-      if (!src_ref) return vil3d_image_view_base_sptr(); \
-      vil3d_convert_stretch_range(src_ref, inter, lo, hi); \
-      vil3d_convert_cast(inter, dest_ref); \
-      break; } 
-macro(VIL_PIXEL_FORMAT_BYTE, vxl_byte )
-macro(VIL_PIXEL_FORMAT_SBYTE , vxl_sbyte )
-macro(VIL_PIXEL_FORMAT_UINT_32 , vxl_uint_32 )
-macro(VIL_PIXEL_FORMAT_UINT_16 , vxl_uint_16 )
-macro(VIL_PIXEL_FORMAT_INT_32 , vxl_int_32 )
-macro(VIL_PIXEL_FORMAT_INT_16 , vxl_int_16 )
-macro(VIL_PIXEL_FORMAT_FLOAT , float )
-macro(VIL_PIXEL_FORMAT_DOUBLE , double )
+   case F: { \
+    vil3d_image_view<T> src_ref = src; \
+    if (!src_ref) return vil3d_image_view_base_sptr(); \
+    vil3d_convert_stretch_range(src_ref, inter, lo, hi); \
+    vil3d_convert_cast(inter, dest_ref); \
+    break; }
+    macro(VIL_PIXEL_FORMAT_BYTE, vxl_byte )
+    macro(VIL_PIXEL_FORMAT_SBYTE , vxl_sbyte )
+    macro(VIL_PIXEL_FORMAT_UINT_32 , vxl_uint_32 )
+    macro(VIL_PIXEL_FORMAT_UINT_16 , vxl_uint_16 )
+    macro(VIL_PIXEL_FORMAT_INT_32 , vxl_int_32 )
+    macro(VIL_PIXEL_FORMAT_INT_16 , vxl_int_16 )
+    macro(VIL_PIXEL_FORMAT_FLOAT , float )
+    macro(VIL_PIXEL_FORMAT_DOUBLE , double )
 #undef macro
-    default:
-      dest_ref.clear();
+   default:
+    dest_ref.clear();
   }
   return dest;
 }
